Move sequencer
