+++ cmd/group-joiner/main.go
@@ -42,7 +42,7 @@ type handler struct {
 	lastBatchId     int
 	eofReceived     int
 	totalPartitions int
-	sequencer       map[int]*utils.Sequencer
+	sequencer       map[int]*middleware.Sequencer
 }
 
 func buildHandler(partition int) middleware.HandlerFunc[*handler] {
@@ -122,9 +122,9 @@ func main() {
 
 	nodeCfg := middleware.Config[*handler]{
 		Builder: func(clientID int) *handler {
-			sequencer := make(map[int]*utils.Sequencer)
+			sequencer := make(map[int]*middleware.Sequencer)
 			for i := 1; i <= cfg.Partitions; i++ {
-				sequencer[i] = utils.NewSequencer()
+				sequencer[i] = middleware.NewSequencer()
 			}
 			return &handler{
 				output:          cfg.Output,
