+++ middleware/filter.go
@@ -1,8 +1,6 @@
 package middleware
 
 import (
-	"distribuidos/tp1/utils"
-
 	amqp "github.com/rabbitmq/amqp091-go"
 )
 
@@ -25,7 +23,7 @@ type filterHandler[T any] struct {
 	filter     FilterFunc[T]
 	partitions map[string]Batch[T]
 	stats      map[string]int
-	sequencer  *utils.Sequencer
+	sequencer  *Sequencer
 }
 
 func (h *filterHandler[T]) handle(ch *Channel, data []byte) error {
@@ -116,7 +114,7 @@ func NewFilter[T any](config FilterConfig, f FilterFunc[T]) (*Node[*filterHandle
 				filter:     f,
 				partitions: partitions,
 				stats:      make(map[string]int),
-				sequencer:  utils.NewSequencer(),
+				sequencer:  NewSequencer(),
 			}
 		},
 		Endpoints: map[string]HandlerFunc[*filterHandler[T]]{
